requirements.txt
asgiref==3.4.1
Django==3.2.8
django-ninja==0.16.1
pydantic==1.8.2
python-frontmatter==1.0.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
